Rearranged the class
